make just a classmethod of BuildsFn
