+++ src/hydra_zen/structured_configs/_implementations.py
@@ -4,6 +4,7 @@ import functools
 import inspect
 import sys
 import warnings
+from collections import Counter, deque
 from dataclasses import (  # use this for runtime checks
     MISSING,
     Field as _Field,
@@ -15,12 +16,14 @@ from dataclasses import (  # use this for runtime checks
     make_dataclass,
 )
 from enum import Enum
+from functools import partial
 from itertools import chain
-from pathlib import Path
+from pathlib import Path, PosixPath, WindowsPath
 from typing import (
     Any,
     Callable,
     Dict,
+    FrozenSet,
     Generic,
     List,
     Mapping,
@@ -78,13 +81,19 @@ from hydra_zen.typing import (
 from hydra_zen.typing._implementations import (
     AllConvert,
     AnyBuilds,
+    Builds,
     BuildsWithSig,
     DataClass_,
+    DataclassOptions,
     DefaultsList,
     Field,
     HasTarget,
     InstOrType,
+    Just as JustT,
+    Partial,
     ZenConvert,
+    _HydraPrimitive,
+    _SupportedViaBuilds,
 )
 
 from ._globals import (
@@ -109,13 +118,22 @@ from ._type_guards import (
     safe_getattr,
     uses_zen_processing,
 )
+from ._utils import merge_settings
 
 T = TypeVar("T")
 _T = TypeVar("_T")
-Importable = TypeVar("Importable", bound=Callable[..., Any])
 P = ParamSpec("P")
 R = TypeVar("R")
+
+TD = TypeVar("TD", bound=DataClass_)
+TC = TypeVar("TC", bound=Callable[..., Any])
+TP = TypeVar("TP", bound=_HydraPrimitive)
+TB = TypeVar("TB", bound=Union[_SupportedViaBuilds, FrozenSet[Any]])
+
+Importable = TypeVar("Importable", bound=Callable[..., Any])
 Field_Entry: TypeAlias = Tuple[str, type, Field[Any]]
+
+_JUST_CONVERT_SETTINGS = AllConvert(dataclass=True, flat_target=False)
 
 # default zen_convert settings for `builds` and `hydrated_dataclass`
 _BUILDS_CONVERT_SETTINGS = AllConvert(dataclass=True, flat_target=True)
@@ -1022,6 +1040,26 @@ class BuildsFn(Generic[T]):
     ) -> type:
         """Broadens a type annotation until it is compatible with Hydra.
 
+        Override this to change how `builds` refines the type annotations
+        of the configs that it produces.
+
+        Parameters
+        ----------
+        type_ : Any
+            The type being sanitized.
+
+        primitive_only: bool, optional (default=False)
+            If true, only `bool` | `None` | `int` | `float` | str`
+            is permitted.
+
+        wrap_optional: bool, optional (default=False)
+            `True` indicates that the resulting type should be
+            wrapped in `Optional`.
+
+        nested: bool, optional (default=False)
+            `True` indicates that this function is processing a type within
+            a container type.
+
         Examples
         --------
         >>> sanitized_type(int)
@@ -1194,7 +1232,10 @@ class BuildsFn(Generic[T]):
 
     @classmethod
     def _get_obj_path(cls, target: Any) -> str:
-        """Used to get the `_target_` value for the resulting config."""
+        """Used to get the `_target_` value for the resulting config.
+
+        Override this to control how `builds` determines the _target_ field
+        in the configs that it produces."""
         name = _utils.safe_name(target, repr_allowed=False)
 
         if name == _utils.UNKNOWN_NAME:
@@ -1243,7 +1284,7 @@ class BuildsFn(Generic[T]):
         return Just(path=cls._get_obj_path(obj))
 
     @classmethod
-    def mutable_value(cls, x: _T, *, zen_convert: Optional[ZenConvert] = None) -> _T:
+    def _mutable_value(cls, x: _T, *, zen_convert: Optional[ZenConvert] = None) -> _T:
         """Used to set a mutable object as a default value for a field
         in a dataclass.
 
@@ -1279,12 +1320,12 @@ class BuildsFn(Generic[T]):
         del zen_convert
 
         if cast in {list, tuple, dict}:
-            x = cls.sanitize_collection(x, convert_dataclass=settings["dataclass"])
+            x = cls._sanitize_collection(x, convert_dataclass=settings["dataclass"])
             return field(default_factory=lambda: cast(x))
         return field(default_factory=lambda: x)
 
     @classmethod
-    def sanitized_default_value(
+    def _sanitized_default_value(
         cls,
         value: Any,
         allow_zen_conversion: bool = True,
@@ -1298,7 +1339,9 @@ class BuildsFn(Generic[T]):
         zen_dataclass: Optional[DataclassOptions] = None,
     ) -> Any:
         """Converts `value` to Hydra-supported type if necessary and possible. Otherwise
-        raises `HydraZenUnsupportedPrimitiveError`"""
+        raises `HydraZenUnsupportedPrimitiveError`.
+
+        Override this method to add"""
         # Common primitives supported by Hydra.
         # We check exhaustively for all Hydra-supported primitives below but seek to
         # speedup checks for common types here.
@@ -1307,7 +1350,7 @@ class BuildsFn(Generic[T]):
 
         # non-str collection
         if hasattr(value, "__iter__"):
-            value = cls.sanitize_collection(
+            value = cls._sanitize_collection(
                 value,
                 convert_dataclass=convert_dataclass,
                 hydra_convert=hydra_convert,
@@ -1340,7 +1383,7 @@ class BuildsFn(Generic[T]):
             for _field in _val_fields:
                 if _field.init and hasattr(value, _field.name):
                     _val = safe_getattr(value, _field.name)
-                    converted_fields[_field.name] = cls.sanitized_default_value(
+                    converted_fields[_field.name] = cls._sanitized_default_value(
                         _val,
                         allow_zen_conversion=allow_zen_conversion,
                         field_name=_field.name,
@@ -1422,7 +1465,7 @@ class BuildsFn(Generic[T]):
             if isinstance(_val, pydantic.fields.UndefinedType):
                 return MISSING
 
-            return cls.sanitized_default_value(
+            return cls._sanitized_default_value(
                 _val,
                 allow_zen_conversion=allow_zen_conversion,
                 error_prefix=error_prefix,
@@ -1467,7 +1510,7 @@ class BuildsFn(Generic[T]):
         raise HydraZenUnsupportedPrimitiveError(err_msg)
 
     @classmethod
-    def sanitize_collection(
+    def _sanitize_collection(
         cls,
         x: _T,
         *,
@@ -1479,7 +1522,7 @@ class BuildsFn(Generic[T]):
         type_x = type(x)
         if type_x in {list, tuple}:
             return type_x(
-                cls.sanitized_default_value(
+                cls._sanitized_default_value(
                     _x,
                     convert_dataclass=convert_dataclass,
                     hydra_convert=hydra_convert,
@@ -1491,13 +1534,13 @@ class BuildsFn(Generic[T]):
             return {
                 # Hydra doesn't permit structured configs for keys, thus we only
                 # support its basic primitives here.
-                cls.sanitized_default_value(
+                cls._sanitized_default_value(
                     k,
                     allow_zen_conversion=False,
                     structured_conf_permitted=False,
                     error_prefix="Configuring dictionary key:",
                     convert_dataclass=False,
-                ): cls.sanitized_default_value(
+                ): cls._sanitized_default_value(
                     v,
                     convert_dataclass=convert_dataclass,
                     hydra_convert=hydra_convert,
@@ -1510,7 +1553,7 @@ class BuildsFn(Generic[T]):
             return x
 
     @classmethod
-    def sanitized_field(
+    def _sanitized_field(
         cls,
         value: Any,
         init: bool = True,
@@ -1520,7 +1563,7 @@ class BuildsFn(Generic[T]):
         field_name: str = "",
         convert_dataclass: bool,
     ) -> Field[Any]:
-        value = cls.sanitized_default_value(
+        value = cls._sanitized_default_value(
             value,
             allow_zen_conversion=allow_zen_conversion,
             error_prefix=error_prefix,
@@ -2148,7 +2191,7 @@ class BuildsFn(Generic[T]):
                     f"`hydra_defaults` must be type `None | list[str | dict[str, str | list[str] | None ]]`"
                     f", Got: {repr(hydra_defaults)}"
                 )
-            hydra_defaults = self.sanitize_collection(
+            hydra_defaults = self._sanitize_collection(
                 hydra_defaults, convert_dataclass=False
             )
             base_fields.append(
@@ -2169,7 +2212,7 @@ class BuildsFn(Generic[T]):
                     Tuple[Any, ...],
                     _utils.field(
                         default=tuple(
-                            self.sanitized_default_value(
+                            self._sanitized_default_value(
                                 x,
                                 error_prefix=BUILDS_ERROR_PREFIX,
                                 convert_dataclass=zen_convert_settings["dataclass"],
@@ -2304,7 +2347,7 @@ class BuildsFn(Generic[T]):
 
                 # validates
                 _pos_args = tuple(
-                    self.sanitized_default_value(
+                    self._sanitized_default_value(
                         x, allow_zen_conversion=False, convert_dataclass=False
                     )
                     for x in _pos_args
@@ -2512,7 +2555,7 @@ class BuildsFn(Generic[T]):
             for field_ in fields(base):
                 if field_.default is not MISSING:
                     # performs validation
-                    self.sanitized_default_value(
+                    self._sanitized_default_value(
                         field_.default,
                         allow_zen_conversion=False,
                         error_prefix=BUILDS_ERROR_PREFIX,
@@ -2536,7 +2579,7 @@ class BuildsFn(Generic[T]):
                 value = item[-1]
 
                 if not isinstance(value, _Field):
-                    _field = self.sanitized_field(
+                    _field = self._sanitized_field(
                         value,
                         error_prefix=BUILDS_ERROR_PREFIX,
                         field_name=item[0],
@@ -2594,8 +2637,326 @@ class BuildsFn(Generic[T]):
             Union[Type[Builds[Importable]], Type[BuildsWithSig[Type[R], P]]], out
         )
 
+    @overload
+    @classmethod
+    def just(
+        cls,
+        obj: TP,
+        *,
+        zen_convert: Optional[ZenConvert] = ...,
+        hydra_recursive: Optional[bool] = ...,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = ...,
+        zen_dataclass: Optional[DataclassOptions] = ...,
+    ) -> TP:
+        ...
+
+    @overload
+    @classmethod
+    def just(
+        cls,
+        obj: complex,
+        *,
+        zen_convert: Optional[ZenConvert] = ...,
+        hydra_recursive: Optional[bool] = ...,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = ...,
+        zen_dataclass: Optional[DataclassOptions] = ...,
+    ) -> "ConfigComplex":
+        ...
+
+    @overload
+    @classmethod
+    def just(
+        cls,
+        obj: TC,
+        *,
+        zen_convert: Optional[ZenConvert] = ...,
+        hydra_recursive: Optional[bool] = ...,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = ...,
+        zen_dataclass: Optional[DataclassOptions] = ...,
+    ) -> JustT[TC]:
+        ...
+
+    @overload
+    @classmethod
+    def just(
+        cls,
+        obj: TB,
+        *,
+        zen_convert: Optional[ZenConvert] = ...,
+        hydra_recursive: Optional[bool] = ...,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = ...,
+        zen_dataclass: Optional[DataclassOptions] = ...,
+    ) -> Builds[Type[TB]]:
+        ...
+
+    @overload
+    @classmethod
+    def just(
+        cls,
+        obj: TD,
+        *,
+        zen_convert: Literal[None] = ...,
+        hydra_recursive: Optional[bool] = ...,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = ...,
+        zen_dataclass: Optional[DataclassOptions] = ...,
+    ) -> Type[Builds[Type[TD]]]:
+        ...
+
+    @overload
+    @classmethod
+    def just(
+        cls,
+        obj: DataClass_,
+        *,
+        zen_convert: ZenConvert,
+        hydra_recursive: Optional[bool] = ...,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = ...,
+        zen_dataclass: Optional[DataclassOptions] = ...,
+    ) -> Any:
+        ...
+
+    @overload
+    @classmethod
+    def just(
+        cls,
+        obj: Any,
+        *,
+        zen_convert: Optional[ZenConvert] = ...,
+        hydra_recursive: Optional[bool] = ...,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = ...,
+        zen_dataclass: Optional[DataclassOptions] = ...,
+    ) -> Any:
+        ...
+
+    @classmethod
+    def just(
+        cls,
+        obj: Any,
+        *,
+        zen_convert: Optional[ZenConvert] = None,
+        hydra_recursive: Optional[bool] = None,
+        hydra_convert: Optional[Literal["none", "partial", "all", "object"]] = None,
+        zen_dataclass: Optional[DataclassOptions] = None,
+    ) -> Any:
+        """`just(obj)` returns a config that, when instantiated, just returns `obj`.
+
+        I.e., `instantiate(just(obj)) == obj`
+
+        `just` is designed to be idempotent. I.e., `just(obj) == just(just(obj))`
+
+        Parameters
+        ----------
+        obj : Callable[..., Any] | HydraSupportedPrimitive | ZenSupportedPrimitive
+            A type (e.g. a class-object), function-object, or a value that is either
+            supported by Hydra or has auto-config support via hydra-zen.
+
+        zen_convert : Optional[ZenConvert]
+            A dictionary that modifies hydra-zen's value and type conversion behavior.
+            Consists of the following optional key-value pairs (:ref:`zen-convert`):
+
+            - `dataclass` : `bool` (default=True):
+                If `True` any dataclass type/instance without a
+                `_target_` field is automatically converted to a targeted config
+                that will instantiate to that type/instance. Otherwise the dataclass
+                type/instance will be passed through as-is.
+
+
+        hydra_recursive : Optional[bool], optional (default=True)
+            If ``True``, then Hydra will recursively instantiate all other
+            hydra-config objects nested within this config [2]_.
+
+            If ``None``, the ``_recursive_`` attribute is not set on the resulting config.
+
+        hydra_convert : Optional[Literal["none", "partial", "all", "object"]], optional (default="none")
+            Determines how Hydra treats the non-primitive, omegaconf-specific objects
+            during instantiateion [3]_.
+
+            - ``"none"``: No conversion occurs; omegaconf containers are passed through (Default)
+            - ``"partial"``: ``DictConfig`` and ``ListConfig`` objects converted to ``dict`` and
+            ``list``, respectively. Structured configs and their fields are passed without conversion.
+            - ``"all"``: All passed objects are converted to dicts, lists, and primitives, without
+            a trace of OmegaConf containers.
+            - ``"object"``: Passed objects are converted to dict and list. Structured Configs are converted to instances of the backing dataclass / attr class.
+
+            If ``None``, the ``_convert_`` attribute is not set on the resulting config.
+
+        zen_dataclass : Optional[DataclassOptions]
+            A dictionary that can specify any option that is supported by
+            :py:func:`dataclasses.make_dataclass` other than `fields`.
+            The default value for `unsafe_hash` is `True`.
+
+            These options are only relevant when the input to `just` is a dataclass
+            instance. Otherwise, `just` does not utilize these options when auto-generating
+            configs.
+
+            Additionally, the `module` field can be specified to enable pickle
+            compatibility. See `hydra_zen.typing.DataclassOptions` for details.
+
+        Returns
+        -------
+        out : HydraSupportedPrimitive | Builds[Type[obj]]
+            ``out`` is ``obj`` unchanged if ``obj`` is supported natively by Hydra, otherwise ``out`` is a dynamically-generated dataclass type or instance.
+
+        Raises
+        ------
+        hydra_zen.errors.HydraZenUnsupportedPrimitiveError
+
+        References
+        ----------
+        .. [1] https://hydra.cc/docs/tutorials/structured_config/intro/
+        .. [2] https://hydra.cc/docs/advanced/instantiate_objects/overview/#recursive-instantiation
+        .. [3] https://hydra.cc/docs/advanced/instantiate_objects/overview/#parameter-conversion-strategies
+
+        See Also
+        --------
+        builds : Create a targeted structured config designed to "build" a particular object.
+        make_config: Creates a general config with customized field names, default values, and annotations.
+
+        Notes
+        -----
+        Here, a "config" is a dynamically-generated dataclass type that is designed to be
+        compatible with Hydra [1]_.
+
+        The configs produced by ``just(<type_or_func>)`` introduce an explicit dependency
+        on hydra-zen. I.e. hydra-zen must be installed in order to instantiate any config
+        that used `just`.
+
+        Examples
+        --------
+        **Basic usage**
+
+        >>> from hydra_zen import just, instantiate, to_yaml
+
+        `just`, called on a value of a type that is natively supported by Hydra,
+        will return that value unchanged:
+
+        >>> just(1)
+        1
+        >>> just({"a": False})
+        {"a": False}
+
+        `just` can be used to create a config that will simply import a class-object or function when the config is instantiated by Hydra.
+
+
+        >>> class A: pass
+        >>> def my_func(x: int): pass
+
+        >>> instantiate(just(A)) is A
+        True
+        >>> instantiate(just(my_func)) is my_func
+        True
+
+        `just` dynamically generates dataclass types, a.k.a structured configs, to describe
+        ``obj``
+
+        >>> just(A)
+        Just_A(_target_='hydra_zen.funcs.get_obj', path='__main__.A')
+        >>> just(my_func)
+        Just_my_func(_target_='hydra_zen.funcs.get_obj', path='__main__.my_func')
+
+        **Auto-config support**
+
+        Calling `just` on a value of a type that has
+        :ref:`special support from hydra-zen <additional-types>`
+        will return a structured config instance that, when instantiated by Hydra, returns
+        that value.
+
+        >>> just(1+2j)
+        ConfigComplex(real=1.0, imag=2.0, _target_='builtins.complex')
+        >>> instantiate(just(1+2j))
+        (1+2j)
+
+        >>> just({1, 2, 3})
+        Builds_set(_target_='builtins.set', _args_=((1, 2, 3),))
+        >>> instantiate(just({1, 2, 3}))
+        {1, 2, 3}
+
+        By default, `just` will convert a dataclass instance to a corresponding targeted
+        config. This behavior can be modified via :ref:`zen-convert settings<zen-convert>`.
+
+        >>> from dataclasses import dataclass
+        >>> @dataclass
+        ... class B:
+        ...     x: complex
+        >>>
+        >>> instantiate(just(B)) is B
+        True
+        >>> instantiate(just(B(2+3j))) == B(2+3j)
+        True
+
+        `just` operates recursively within sequences, mappings, and dataclass fields.
+
+        >>> z_dict = {'a': [3-4j, 1+2j]}
+        >>> assert instantiate(just(z_dict)) == z_dict
+        >>>
+        >>> from typing import Any
+        >>> @dataclass
+        ... class C:
+        ...     x: Any
+        >>>
+        >>> instantiate(just(C(B(2+3j))))
+        C(B(2+3j))
+
+        **Auto-Application of just**
+
+        Both `builds` and `make_config` will automatically (and recursively) apply
+        `just` to all configured values. E.g. in the following example `just` will be
+        applied to both  the complex-valued the list and to ``sum``.
+
+        >>> from hydra_zen import make_config
+        >>> Conf2 = make_config(data=[1+2j, 2+3j], reduction_fn=sum)
+
+        >>> print(to_yaml(Conf2))
+        data:
+        - real: 1.0
+        imag: 2.0
+        _target_: builtins.complex
+        - real: 2.0
+        imag: 3.0
+        _target_: builtins.complex
+        reduction_fn:
+        _target_: hydra_zen.funcs.get_obj
+        path: builtins.sum
+
+        >>> conf = instantiate(Conf2)
+        >>> conf.reduction_fn(conf.data)
+        (3+5j)
+        """
+        convert_settings = merge_settings(zen_convert, _JUST_CONVERT_SETTINGS)
+        del zen_convert
+        _utils.validate_hydra_options(
+            hydra_recursive=hydra_recursive, hydra_convert=hydra_convert
+        )
+        if zen_dataclass is None:
+            zen_dataclass = {}
+
+        return cls._sanitized_default_value(
+            obj,
+            allow_zen_conversion=True,
+            structured_conf_permitted=True,
+            field_name="",
+            error_prefix="",
+            convert_dataclass=convert_settings["dataclass"],
+            hydra_convert=hydra_convert,
+            hydra_recursive=hydra_recursive,
+            zen_dataclass=_utils.parse_dataclass_options(zen_dataclass),
+        )
+
 
 builds: BuildsFn[SupportedPrimitive] = BuildsFn[SupportedPrimitive]("builds")
+
+
+@dataclass(unsafe_hash=True)
+class ConfigComplex:
+    real: Any
+    imag: Any
+    _target_: str = field(default=builds._get_obj_path(complex), init=False)
+
+
+@dataclass(unsafe_hash=True)
+class ConfigPath:
+    _args_: Tuple[str]
+    _target_: str = field(default=builds._get_obj_path(Path), init=False)
 
 
 @overload
@@ -2729,4 +3090,95 @@ def mutable_value(
     HasMutableDefault(a_list=[1, 2, 3, -1])
     >>> HasMutableDefault()
     HasMutableDefault(a_list=[1, 2, 3])"""
-    return BuildsFunction.mutable_value(x, zen_convert=zen_convert)
+    return BuildsFunction._mutable_value(x, zen_convert=zen_convert)
+
+
+def convert_complex(value: complex) -> Builds[Type[complex]]:
+    return cast(Builds[Type[complex]], ConfigComplex(real=value.real, imag=value.imag))
+
+
+ZEN_VALUE_CONVERSION[complex] = convert_complex
+
+
+if Path in ZEN_SUPPORTED_PRIMITIVES:  # pragma: no cover
+
+    def convert_path(value: Path) -> Builds[Type[Path]]:
+        return cast(Builds[Type[Path]], ConfigPath(_args_=(str(value),)))
+
+    ZEN_VALUE_CONVERSION[Path] = convert_path
+    ZEN_VALUE_CONVERSION[PosixPath] = convert_path
+    ZEN_VALUE_CONVERSION[WindowsPath] = convert_path
+
+
+def _unpack_partial(value: Partial[_T]) -> PartialBuilds[Type[_T]]:
+    target = cast(Type[_T], value.func)
+    return builds(target, *value.args, **value.keywords, zen_partial=True)()
+
+
+@dataclass(unsafe_hash=True)
+class ConfigFromTuple:
+    _args_: Tuple[Any, ...]
+    _target_: str
+
+    def __post_init__(self):
+        self._args_ = (
+            builds._sanitized_default_value(
+                tuple(self._args_),
+                convert_dataclass=True,
+                allow_zen_conversion=True,
+                structured_conf_permitted=True,
+            ),
+        )
+
+
+@dataclass(unsafe_hash=True)
+class ConfigFromDict:
+    _args_: Any
+    _target_: str
+
+    def __post_init__(self):
+        self._args_ = (
+            builds._sanitized_default_value(
+                dict(self._args_),
+                convert_dataclass=True,
+                allow_zen_conversion=True,
+                structured_conf_permitted=True,
+            ),
+        )
+
+
+@dataclass(unsafe_hash=True)
+class ConfigRange:
+    start: InitVar[int]
+    stop: InitVar[int]
+    step: InitVar[int]
+    _target_: str = field(default=builds._get_obj_path(range), init=False)
+    _args_: Tuple[int, ...] = field(default=(), init=False, repr=False)
+
+    def __post_init__(self, start, stop, step):
+        self._args_ = (start, stop, step)
+
+
+ZEN_VALUE_CONVERSION[set] = partial(ConfigFromTuple, _target_=builds._get_obj_path(set))
+ZEN_VALUE_CONVERSION[frozenset] = partial(
+    ConfigFromTuple, _target_=builds._get_obj_path(frozenset)
+)
+ZEN_VALUE_CONVERSION[deque] = partial(
+    ConfigFromTuple, _target_=builds._get_obj_path(deque)
+)
+
+if bytes in ZEN_SUPPORTED_PRIMITIVES:  # pragma: no cover
+    ZEN_VALUE_CONVERSION[bytes] = partial(
+        ConfigFromTuple, _target_=builds._get_obj_path(bytes)
+    )
+
+ZEN_VALUE_CONVERSION[bytearray] = partial(
+    ConfigFromTuple, _target_=builds._get_obj_path(bytearray)
+)
+ZEN_VALUE_CONVERSION[range] = lambda value: ConfigRange(
+    value.start, value.stop, value.step
+)
+ZEN_VALUE_CONVERSION[Counter] = partial(
+    ConfigFromDict, _target_=builds._get_obj_path(Counter)
+)
+ZEN_VALUE_CONVERSION[functools.partial] = _unpack_partial
